+++ articles/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ArticlesConfig(AppConfig):
+    name = 'articles'
